restructured to pattern MVVM
