+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         android:theme="@style/Theme.QuixaFood"
         tools:targetApi="31">
 
-        <receiver android:name=".nofications.AlarmReceiver" android:exported="true" />
+        <receiver android:name=".model.nofications.AlarmReceiver" android:exported="true" />
         <activity
             android:name=".MainActivity"
             android:exported="true"
