+++ app/src/main/java/com/example/quixafood/model/navigation/NavGraph.kt
@@ -1,8 +1,7 @@
-package com.example.quixafood.navigation
+package com.example.quixafood.model.navigation
 
 import android.content.Context
 import android.os.Build
-import android.widget.Toast
 import androidx.annotation.RequiresApi
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
@@ -24,17 +23,17 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import com.example.quixafood.data.AuthRepository
-import com.example.quixafood.models.mockItens
-import com.example.quixafood.ui.components.BottomNavigationBar
-import com.example.quixafood.ui.screens.DetailsScreen
-import com.example.quixafood.ui.screens.FavoritesScreen
-import com.example.quixafood.ui.screens.HelpScreen
-import com.example.quixafood.ui.screens.HomeScreen
-import com.example.quixafood.ui.screens.LoginScreen
-import com.example.quixafood.ui.screens.RegisterScreen
-import com.example.quixafood.ui.screens.ResetPasswordScreen
-import com.example.quixafood.ui.screens.SearchScreen
-import com.example.quixafood.ui.screens.SettingsScreen
+import com.example.quixafood.model.itemmodel.mockItens
+import com.example.quixafood.ui.view.components.BottomNavigationBar
+import com.example.quixafood.ui.view.screens.DetailsScreen
+import com.example.quixafood.ui.view.screens.FavoritesScreen
+import com.example.quixafood.ui.view.screens.HelpScreen
+import com.example.quixafood.ui.view.screens.HomeScreen
+import com.example.quixafood.ui.view.screens.LoginScreen
+import com.example.quixafood.ui.view.screens.RegisterScreen
+import com.example.quixafood.ui.view.screens.ResetPasswordScreen
+import com.example.quixafood.ui.view.screens.SearchScreen
+import com.example.quixafood.ui.view.screens.SettingsScreen
 import com.example.quixafood.ui.theme.QuixaFoodTheme
 import com.example.quixafood.viewmodel.AuthViewModel
 import java.time.LocalTime
