+++ app/src/main/java/com/example/quixafood/ui/view/components/ItemDetailsActivity.kt
@@ -1,0 +1,1 @@
+package com.example.quixafood.ui.view.components
