+++ app/src/main/java/com/example/quixafood/ui/view/components/PacManProgressBar.kt
@@ -1,4 +1,4 @@
-package com.example.quixafood.ui.components
+package com.example.quixafood.ui.view.components
 
 
 import androidx.compose.animation.core.*
