+++ app/src/main/java/com/example/quixafood/ui/view/screens/RegisterScreen.kt
@@ -1,4 +1,4 @@
-package com.example.quixafood.ui.screens
+package com.example.quixafood.ui.view.screens
 
 import android.content.Context
 import android.util.Log
